requirements.txt
fastapi>=0.135.2
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
pydantic>=2.9.0
pydantic-settings>=2.0.0
PyJWT>=2.8.0
python-multipart>=0.0.9
alembic>=1.13.0
httpx>=0.27.0
google-genai